+++ src/map-director/slopes-group-filler/OutRampTilesetTile.hpp
@@ -24,31 +24,9 @@
 
 class OutRampPropertiesLoader final : public RampPropertiesLoaderBase {
 private:
-    TileCornerElevations elevation_offsets_for
-        (CardinalDirection direction) const final
-    {
-        using Cd = CardinalDirection;
-        switch (direction) {
-        case Cd::north_east: return TileCornerElevations{0, 0, 1, 0};
-        case Cd::north_west: return TileCornerElevations{0, 0, 0, 1};
-        case Cd::south_east: return TileCornerElevations{0, 1, 0, 0};
-        case Cd::south_west: return TileCornerElevations{1, 0, 0, 0};
-        default:
-            throw InvalidArgument{"direction bad"};
-        }
-    }
+    TileCornerElevations elevation_offsets_for(CardinalDirection) const final;
 
-    Orientation orientation_for(CardinalDirection direction) const final {
-        using Cd = CardinalDirection;
-        switch (direction) {
-        case Cd::north_east: case Cd::south_west:
-            return Orientation::nw_to_se_elements;
-        case Cd::north_west: case Cd::south_east:
-            return Orientation::sw_to_ne_elements;
-        default:
-            throw InvalidArgument{"direction bad"};
-        }
-    }
+    Orientation orientation_for(CardinalDirection) const final;
 };
 
 // ----------------------------------------------------------------------------
