+++ src/map-director/slopes-group-filler/TwoWayWallSplits.cpp
@@ -18,9 +18,7 @@
 
 *****************************************************************************/
 
-#include "TwoWaySplit.hpp"
-#include "OutWallTilesetTile.hpp"
-#include "InWallCornerSplits.hpp"
+#include "TwoWayWallSplits.hpp"
 
 #include "../../TriangleSegment.hpp"
 
@@ -29,13 +27,13 @@ namespace {
 using Triangle = TriangleSegment;
 
 class NorthGenerationStrategy final :
-    public TwoWaySplit::GeometryGenerationStrategy
+    public SplitWallGeometry::GeometryGenerationStrategy
 {
 public:
     void with_splitter_do
         (const TileCornerElevations & elevations,
          Real division_z,
-         const TwoWaySplit::WithTwoWaySplit & with_split_callback) const final
+         const SplitWallGeometry::WithSplitWallGeometry & with_split_callback) const final
     {
         NorthSouthSplit nss{elevations, division_z};
         with_split_callback(nss);
@@ -53,13 +51,13 @@ public:
 };
 
 class SouthGenerationStrategy final :
-    public TwoWaySplit::GeometryGenerationStrategy
+    public SplitWallGeometry::GeometryGenerationStrategy
 {
 public:
     void with_splitter_do
         (const TileCornerElevations & elevations,
          Real division_z,
-         const TwoWaySplit::WithTwoWaySplit & with_split_callback) const final
+         const SplitWallGeometry::WithSplitWallGeometry & with_split_callback) const final
     {
         SouthNorthSplit sns{elevations, division_z};
         with_split_callback(sns);
@@ -77,13 +75,13 @@ public:
 };
 
 class EastGenerationStrategy final :
-    public TwoWaySplit::GeometryGenerationStrategy
+    public SplitWallGeometry::GeometryGenerationStrategy
 {
 public:
     void with_splitter_do
         (const TileCornerElevations & elevations,
          Real division_z,
-         const TwoWaySplit::WithTwoWaySplit & with_split_callback) const final
+         const SplitWallGeometry::WithSplitWallGeometry & with_split_callback) const final
     {
         EastWestSplit ews{elevations, division_z};
         with_split_callback(ews);
@@ -101,13 +99,13 @@ public:
 };
 
 class WestGenerationStrategy final :
-    public TwoWaySplit::GeometryGenerationStrategy
+    public SplitWallGeometry::GeometryGenerationStrategy
 {
 public:
     void with_splitter_do
         (const TileCornerElevations & elevations,
          Real division_z,
-         const TwoWaySplit::WithTwoWaySplit & with_split_callback) const final
+         const SplitWallGeometry::WithSplitWallGeometry & with_split_callback) const final
     {
         WestEastSplit wes{elevations, division_z};
         with_split_callback(wes);
@@ -124,149 +122,12 @@ public:
     }
 };
 
-StripVertex::StripSide other_side_of(StripVertex::StripSide side) {
-    using Side = StripVertex::StripSide;
-    switch (side) {
-    case Side::a: return Side::b;
-    case Side::b: return Side::a;
-    default: break;
-    }
-    throw InvalidArgument{"bad argument"};
-}
-
 } // end of <anonymous> namespace
-
-StripTriangle::StripTriangle
-    (const StripVertex & a,
-     const StripVertex & b,
-     const StripVertex & c):
-    m_a(a),
-    m_b(b),
-    m_c(c) {}
-
-TriangleSegment StripTriangle::to_triangle_segment() const {
-    return TriangleSegment{m_a.point, m_b.point, m_c.point};
-}
-
-StripVertex StripTriangle::vertex_a() const { return m_a; }
-
-StripVertex StripTriangle::vertex_b() const { return m_b; }
-
-StripVertex StripTriangle::vertex_c() const { return m_c; }
-
-/* private */ void LinearStripTriangleCollection::triangle_strip
-    (const Vector & a_point,
-     const Vector & b_start,
-     const Vector & b_last,
-     StripVertex::StripSide a_side,
-     int steps_count)
-{
-    if (are_very_close(b_start, b_last)) {
-        throw InvalidArgument{""};
-    }
-    auto b_side_pt = [b_start, b_last] (Real t)
-        { return b_start*(1 - t) + b_last*t; };
-    auto b_side = other_side_of(a_side);
-    for (int i = 0; i != steps_count; ++i) {
-        Real t = Real(i) / Real(steps_count);
-        Real next_t = Real(i + 1) / Real(steps_count);
-        add_triangle(StripTriangle
-            {StripVertex{a_point, {}, a_side},
-             StripVertex{b_side_pt(t     ), t     , b_side},
-             StripVertex{b_side_pt(next_t), next_t, b_side}});
-    }
-}
-
-void LinearStripTriangleCollection::make_strip
-    (const Vector & a_start, const Vector & a_last,
-     const Vector & b_start, const Vector & b_last,
-     int steps_count)
-{
-    using StripSide = StripVertex::StripSide;
-    using Triangle = TriangleSegment;
-    if (steps_count < 0) {
-        throw InvalidArgument{"steps_count must be a non-negative integer"};
-    }
-
-    if (   are_very_close(a_start, a_last)
-        && are_very_close(b_start, b_last))
-    {
-        return;
-    }
-    // atempting to generate a one dimensional line
-    else if (   are_very_close(a_start, b_start)
-             && are_very_close(a_last , b_last ))
-    {
-        return;
-    } else if (steps_count == 0) {
-        return;
-    } else if (are_very_close(a_start, a_last)) {
-        return triangle_strip(a_start, b_start, b_last, StripSide::a, steps_count);
-    } else if (are_very_close(b_start, b_last)) {
-        return triangle_strip(b_start, a_start, a_last, StripSide::b, steps_count);
-    }
-
-    auto a_side_pt = [a_start, a_last] (Real t)
-        { return a_start*(1 - t) + a_last*t; };
-    auto b_side_pt = [b_start, b_last] (Real t)
-        { return b_start*(1 - t) + b_last*t; };
-    auto normal_step = [this, &a_side_pt, &b_side_pt] (Real last, Real next) {
-
-        add_triangle(StripTriangle
-            {StripVertex{a_side_pt(last), last, StripSide::a},
-             StripVertex{b_side_pt(last), last, StripSide::b},
-             StripVertex{b_side_pt(next), next, StripSide::b}});
-        add_triangle(StripTriangle
-            {StripVertex{a_side_pt(last), last, StripSide::a},
-             StripVertex{a_side_pt(next), next, StripSide::a},
-             StripVertex{b_side_pt(next), next, StripSide::b}});
-    };
-    auto first_step =
-        [this, &a_side_pt, &b_side_pt, &a_start, &b_start]
-        (Real next)
-    {
-        assert(are_very_close(a_start, b_start));
-        add_triangle(StripTriangle
-            {StripVertex{a_start, 0, {}},
-             StripVertex{a_side_pt(next), next, StripSide::a},
-             StripVertex{b_side_pt(next), next, StripSide::b}});
-    };
-    int step = 0;
-    Real step_size = 1 / Real(steps_count);
-    auto get_last = [step_size, step] { return step*step_size; };
-    auto get_next = [step_size, step] { return (step + 1)*step_size; };
-
-    if (step == 0 && steps_count > 1) {
-        if (are_very_close(a_start, b_start)) {
-            first_step(get_next());
-        } else {
-            // normal first step
-            normal_step(get_last(), get_next());
-        }
-        ++step;
-    }
-
-    for (; step < steps_count - 1; ++step) {
-        normal_step(get_last(), get_next());
-    }
-
-    if (step == 0 && are_very_close(a_start, b_start)) {
-        first_step(get_next());
-    } else if (are_very_close(a_last, b_last)) {
-        auto last = get_last();
-        add_triangle(StripTriangle
-            {StripVertex{a_side_pt(last), last, StripSide::a},
-             StripVertex{b_side_pt(last), last, StripSide::b},
-             StripVertex{a_last, 1, StripSide::both}});
-    } else {
-        normal_step(get_last(), get_next());
-    }
-}
 
 // ----------------------------------------------------------------------------
 
-/* static */ TwoWaySplit::GeometryGenerationStrategy &
-    TwoWaySplit::choose_geometry_strategy
+/* static */ SplitWallGeometry::GeometryGenerationStrategy &
+    NorthSouthSplit::choose_geometry_strategy
     (CardinalDirection direction)
 {
     static NorthGenerationStrategy north_gen_strat;
@@ -282,40 +143,6 @@ void LinearStripTriangleCollection::make_strip
     }
     // should probably return nullptr or something
     // (as this is based on user data)
-    throw InvalidArgument{"bad direction"};
-}
-
-/* static */ TwoWaySplit::GeometryGenerationStrategy &
-    TwoWaySplit::choose_out_wall_strategy(CardinalDirection direction)
-{
-    static NorthWestOutWallGenerationStrategy nw_out_strat;
-    static NorthEastOutWallGenerationStrategy ne_out_strat;
-    static SouthWestOutWallGenerationStrategy sw_out_strat;
-    static SouthEastOutWallGenerationStrategy se_out_strat;
-    switch (direction) {
-    case CardinalDirection::north_west: return nw_out_strat;
-    case CardinalDirection::north_east: return ne_out_strat;
-    case CardinalDirection::south_west: return sw_out_strat;
-    case CardinalDirection::south_east: return se_out_strat;
-    default: break;
-    }
-    throw InvalidArgument{"bad direction"};
-}
-
-/* static */ TwoWaySplit::GeometryGenerationStrategy &
-    TwoWaySplit::choose_in_wall_strategy(CardinalDirection direction)
-{
-    static NorthWestInWallGenerationStrategy nw_in_strat;
-    static NorthEastInWallGenerationStrategy ne_in_strat;
-    static SouthWestInWallGenerationStrategy sw_in_strat;
-    static SouthEastInWallGenerationStrategy se_in_strat;
-    switch (direction) {
-    case CardinalDirection::north_west: return nw_in_strat;
-    case CardinalDirection::north_east: return ne_in_strat;
-    case CardinalDirection::south_west: return sw_in_strat;
-    case CardinalDirection::south_east: return se_in_strat;
-    default: break;
-    }
     throw InvalidArgument{"bad direction"};
 }
